--- packages/service/support/permission/team/auth.ts
@@ -1,1 +1,0 @@
-import { PermissionValueType } from '@fastgpt/global/support/permission/type';
